requirements.txt
rich>=13.0
typer>=0.12
httpx>=0.27
pydantic>=2.0
anthropic>=0.40
python-dotenv>=1.0
